requirements.txt
--index-url https://download.pytorch.org/whl/cu129
torch==2.8.0+cu129
torchvision==0.23.0+cu129
accelerate==1.10.0
bitsandbytes==0.46.1
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
easyocr==1.7.2
einops==0.8.1
filelock==3.13.1
fsspec==2024.6.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.4
numpy==2.1.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
protobuf==6.31.1
psutil==7.0.0
pyclipper==1.3.0.post6
python-bidi==0.6.6
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.6.1
scikit-image==0.25.2
scipy==1.16.1
sentencepiece==0.2.0
setuptools==70.2.0
shapely==2.1.1
sympy==1.13.3
tifffile==2025.6.11
timm==1.0.19
tokenizers==0.22.0
tqdm==4.67.1
transformers==4.56.1
typing_extensions==4.12.2
urllib3==2.5.0